CsomagGep

+++ src/com/bzsomi/IFreighter.java
@@ -2,5 +2,5 @@ package com.bzsomi;
 
 public interface IFreighter {
 
-    public int csomagokSulya();
+    public double csomagokSulya();
 }
